+++ app/src/main/kotlin/io/treehouses/remote/InitialActivity.kt
@@ -37,6 +37,7 @@ class InitialActivity : BaseInitialActivity() {
         instance = this
         setContentView(bind.root)
         requestPermission()
+        adjustFontScale(resources.configuration)
         val toolbar = findViewById<Toolbar>(R.id.toolbar)
         setSupportActionBar(toolbar)
         currentTitle = "Home"
